// src/EasyHRMS_Angular/angularApp/app/models/payroll-categories/payroll-categories.model.ts
export class PayrollCategories {
    id: number;
    categoryName: string;
    description: string;
    type: boolean;
    percentage: string;
    percentageOf: string;
    amount: number;
    status: string;
    isDeduction: boolean;
    isDefault: string;
    taxDeducted: boolean;
    insurationdeducted: string;
    pensiondeducted: string;
    isBasedOnAttandance: string;
    total: string;
    period: string;
    inbuilt: string;
}